version 1.3.1 prototype
Optimized the Appdata update process that was applied whenever the window size was changed.

A new function 'startup' has been added.
Currently, it is planned to change the way the screen looks at startup or the direction in which it runs in a minimized (tray) form.

The 'Replace' field in the regular expression field has been changed to apply even if a single '\' character is used.