1. Update Packaging project. The packaging project includes all the files that are actually deployed to the Microsoft Store
2. JSON-based Settings Storage
The system has been modified to operate configuration files based on JSON files, and key bindings are also working well. However, some settings are not recognized upon reconnection, which is occurring due to complex setting logic. Various tasks are in progress to simplify the code further.

3. Preparing for Version 2.0
Here's the translation in English:
"In version 2.0, the features that users can directly experience are the ability to import and export setting files based on JSON files, and the ability to process multiple Regex patterns in a chain-like manner.