Added DeepCopy to StringifyAsync to prevent InvalidOperationException
The DeepCopy is performed in the StringifyAsync method to address the issue of repeated Add and Remove calls on the regexChains variable during Drag&Drop operations between GridView items. This prevents the 'Collection was modified; enumeration operation may not execute.' error.

Although using await Task.Delay(100) could indirectly ensure thread safety, it is considered a temporary workaround. Therefore, DeepCopy is used despite its resource consumption.

However, the optimal solution would be to link the Swap method during the Drag&Drop operation in the GridView.

In addition, we have updated some Nuget packages to the latest version.