I have identified several issues with using Keybinds individually within a dual NavigationView structure.
1. Instability of dual NavigationView in WinUI3
In testing, even when no specific actions were assigned, simply opening and closing the Pane caused the NavigationView, which should remain in a secondary layer, to shift to a tertiary layer and then revert. This indicates the presence of bugs, which seem to be attributed not to an internal bug of Delete Newline but rather a WinUI3 bug.

2. Redundancy
Upon entering the Keybinds page, information related to the NavigationView is displayed redundantly. By default, not only the Keybind name but also comments regarding the Regex Chain and Keybind details are displayed in multiple layers. Therefore, displaying all of this information within the NavigationView is not appropriate, and it was concluded that this would degrade user experience. As such, Keybinds information is not fully displayed within the NavigationView.

3. Implementation difficulty
This reason also poses a major challenge.